refactored Morton to move InterleaveBits out to be used in Hilber
